deal with consequences of splitting classes up :(
