+++ arcadia_pycolor/__init__.py
@@ -1,7 +1,9 @@
-from arcadia_pycolor import classes, colors, gradients, mpl, palettes, plot
+from arcadia_pycolor import colors, gradients, mpl, palettes, plot, styles
 
-from .classes import *
 from .colors import *
+from .gradient import *
+from .hexcode import *
+from .palette import *
 
 __all__ = [
     "gradients",
@@ -10,8 +12,6 @@ __all__ = [
     "plot",
 ]
 
-classes_all = [name for name in dir(classes) if not name.startswith("_")]
 colors_all = [name for name in dir(colors) if not name.startswith("_")]
 
-__all__.extend(classes_all)
 __all__.extend(colors_all)
