+++ arcadia_pycolor/tests/test_mpl_conversions.py
@@ -1,20 +1,19 @@
 import pytest
 
 from arcadia_pycolor import Gradient, HexCode, Palette
-from arcadia_pycolor.mpl import gradient_to_linear_cmap, palette_to_cmap
 
 
 def test_palette_to_cmap():
     hex_codes = [HexCode("white", "#FFFFFF"), HexCode("black", "#000000")]
-    assert palette_to_cmap(Palette("my_palette", hex_codes)).colors == hex_codes
+    assert Palette("my_palette", hex_codes).to_mpl_cmap().colors == hex_codes
 
 
 def test_palette_name():
     hex_codes = [HexCode("white", "#FFFFFF"), HexCode("black", "#000000")]
     assert (
-        palette_to_cmap(
-            Palette.from_dict("my_palette", {"white": "#FFFFFF", "black": "#000000"})
-        ).colors
+        Palette.from_dict("my_palette", {"white": "#FFFFFF", "black": "#000000"})
+        .to_mpl_cmap()
+        .colors
         == hex_codes
     )
 
@@ -27,13 +26,11 @@ def test_palette_name():
     ],
 )
 def test_gradient_to_linear_cmap(values):
-    grad = gradient_to_linear_cmap(
-        Gradient(
-            "my_gradient",
-            [HexCode("white", "#FFFFFF"), HexCode("black", "#000000")],
-            values,
-        )
-    )
+    grad = Gradient(
+        "my_gradient",
+        [HexCode("white", "#FFFFFF"), HexCode("black", "#000000")],
+        values,
+    ).to_mpl_linear_cmap()
     assert grad(0) == (1.0, 1.0, 1.0, 1.0)
 
 
@@ -46,11 +43,9 @@ def test_gradient_to_linear_cmap(values):
 )
 def test_gradient_name(values):
     name = "my_gradient"
-    grad = gradient_to_linear_cmap(
-        Gradient(
-            name,
-            [HexCode("white", "#FFFFFF"), HexCode("black", "#000000")],
-            values,
-        )
-    )
+    grad = Gradient(
+        name,
+        [HexCode("white", "#FFFFFF"), HexCode("black", "#000000")],
+        values,
+    ).to_mpl_linear_cmap()
     assert grad.name == name
